backend/requirement.txt
asgiref                      
Django                        
django-cors-headers           
djangorestframework           
djangorestframework-simplejwt 
psycopg2-binary               
PyJWT                         
python-dotenv                 
pytz                          
sqlparse                     
tzdata                        